Move start_your_engine to main.py
