+++ src/database/first_start.py
@@ -5,11 +5,11 @@
     :license:   MIT License.
 """
 
-from database import Base
-from database import Session
-from database import DataBaseMySQL, DataBasePostgreSQL, DataBaseSQLite
+from base import Base
+from base import Session
+from base import metadata_obj
+from main import start_your_engine
 from sqlalchemy.engine import Engine
-from sqlalchemy import MetaData
 from sqlalchemy import Table
 from typing import Optional
 # Table classes - Test if it still works if all tables are commented out
@@ -24,108 +24,6 @@ from typing import Optional
 # from pdm_tables import PdmPurchase
 # from pdm_tables import PdmManufacturer
 # from pdm_tables import PdmVendor
-
-
-# https://dataedo.com/kb/data-glossary/what-is-metadata
-# https://www.geeksforgeeks.org/difference-between-data-and-metadata/
-# https://www.geeksforgeeks.org/metadata-in-dbms-and-its-types/
-metadata_obj = MetaData()
-
-
-# https://stackoverflow.com/questions/73887390/handle-multiple-users-login-database-with-sqlalchemy
-def start_your_engine(url_string: str, db_type: Optional[str], split: Optional[str] = ',', **vargs) -> Engine:
-    """
-    Start your chosen engine
-
-    Parameters
-    ----------
-
-    prefered_engine [str] :
-        What sql engine is preferred to use.
-
-    url [str]:
-        url for database. This can be a single string but also a list of parameters (comma-separated).
-        This list contain the following information: drivername; username; password; host; port; database_name.
-        TODO: add dialect as optional value
-
-    split [str]:
-        split string in list (Optional). Default value is ','.
-
-    **vargs:
-        Other parameters are parsed through
-    """
-    url_list = url_string.split(split)
-
-    if (db_type == "mysql") or (db_type == "MySQL"):
-        msql = DataBaseMySQL()
-
-        if len(url_list) == 1:
-            print("Complete url received.")
-            new_url = url_list[0]
-            dialect = None
-        elif len(url_list) == 6:
-            print("Url shall be created")
-            new_url = msql.make_url(url_list[0], url_list[1], url_list[2], url_list[3], url_list[4], int(url_list[5]))
-            dialect = None
-            # if dialect is part of the url...
-        elif len(url_list) == 7:
-            # list including dialect
-            url_dialect = url_list[0] + '+' + url_list[6]
-            new_url = msql.make_url(url_dialect, url_list[1], url_list[2], url_list[3], int(url_list[4]), url_list[5])
-            dialect = url_list[6]
-        else:
-            raise ValueError("{} is not the right amount of values for the url. [1, 6 or 7]\n".format(len(url_list)))
-
-        msql_engine = msql.start_engine(new_url, dialect=dialect, encoding='utf-8', echo=False, future=True)
-        return(msql_engine)
-
-    elif (db_type is None) or (db_type == "postgresql") or (db_type == "PostgresSQL"):
-        psql = DataBasePostgreSQL()
-
-        if len(url_list) == 1:
-            print("Complete url received.")
-            new_url = url_list[0]
-            dialect = None
-        elif len(url_list) == 6:
-            print("Url shall be created")
-            new_url = psql.make_url(url_list[0], url_list[1], url_list[2], url_list[3], url_list[4], int(url_list[5]))
-            dialect = None
-            # if dialect is part of the url...
-        elif len(url_list) == 7:
-            # list including dialect
-            url_dialect = url_list[0] + '+' + url_list[6]
-            new_url = psql.make_url(url_dialect, url_list[1], url_list[2], url_list[3], int(url_list[4]), url_list[5])
-            dialect = url_list[6]
-        else:
-            raise ValueError("{} is not the right amount of values for the url. [1, 6 or 7]\n".format(len(url_list)))
-
-        psql_engine = psql.start_engine(new_url, dialect=dialect, encoding='utf-8', echo=False, future=True)
-        return(psql_engine)  # Not sure if returning this is required
-
-    elif (db_type == "sqlite") or (db_type == "SQLite"):
-        sqli = DataBaseSQLite()
-
-        if len(url_list) == 1:
-            print("Complete url received.")
-            new_url = url_list[0]
-        elif len(url_list) == 6:
-            print("Url shall be created")
-            new_url = sqli.make_url(url_list[0], url_list[1], url_list[2], url_list[3], url_list[4], int(url_list[5]))
-            # dialect = None  # not needed? remove later
-            # if dialect is part of the url...
-        elif len(url_list) == 7:
-            # list including dialect
-            url_dialect = url_list[0] + '+' + url_list[6]
-            new_url = sqli.make_url(url_dialect, url_list[1], url_list[2], url_list[3], int(url_list[4]), url_list[5])
-            # dialect = url_list[6]  # not needed ? remove later
-        else:
-            raise ValueError("{} is not the right amount of values for the url. [1, 6 or 7]\n".format(len(url_list)))
-
-        sqli_engine = sqli.start_engine(new_url, encoding='utf-8', echo=False, future=True)
-        return(sqli_engine)
-
-    else:
-        raise ValueError("{} Is not a Valid input for 'db_type'.".format(db_type))
 
 
 def create_default_tables(engine: Engine):
